store: move user-agent bits into separate module and add tests.
